C exercises from

+++ pointers_arrays_strings/4-rev_array.c
@@ -10,7 +10,7 @@
 void reverse_array(int *a, int n)
 {
 	int i;
-	int temp;
+	long temp;
 
 	for (i = 0; i < (n / 2); i++)
 	{
